requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-flask>=1.3.0
flake8>=7.0.0